refactored into multiple classes following 'Single-Responsibility Principle'
