+++ src/plugin/plugin.php
@@ -10,6 +10,14 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-require 'class-liberation-engine.php';
+require 'class-post-type-manager.php';
+require 'class-meta-fields-manager.php';
+require 'class-rest-api-extender.php';
 
-new Liberation_Engine();
+( function () {
+	$post_type_manager = new Post_Type_Manager();
+	$custom_post_types = $post_type_manager->get_custom_post_types();
+
+	new Meta_Fields_Manager( $custom_post_types );
+	new Rest_API_Extender( $custom_post_types );
+} )();
